+++ Riot/AppDelegate.m
@@ -88,6 +88,8 @@ NSString *const AppDelegateDidValidateEmailNotification = @"AppDelegateDidValida
 NSString *const AppDelegateDidValidateEmailNotificationSIDKey = @"AppDelegateDidValidateEmailNotificationSIDKey";
 NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDelegateDidValidateEmailNotificationClientSecretKey";
 
+NSString *const AppDelegateUniversalLinkDidChangeNotification = @"AppDelegateUniversalLinkDidChangeNotification";
+
 @interface AppDelegate () <PKPushRegistryDelegate, GDPRConsentViewControllerDelegate, KeyVerificationCoordinatorBridgePresenterDelegate, ServiceTermsModalCoordinatorBridgePresenterDelegate>
 {
     /**
@@ -493,6 +495,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
     NSAssert(_masterTabBarController, @"Something wrong in Main.storyboard");
     
     _isAppForeground = NO;
+    _handleSelfVerificationRequest = YES;
     
     // Configure our analytics. It will indeed start if the option is enabled
     [MXSDKOptions sharedInstance].analyticsDelegate = [Analytics sharedInstance];
@@ -2037,6 +2040,19 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
     // iOS Patch: fix vector.im urls before using it
     webURL = [Tools fixURLWithSeveralHashKeys:webURL];
 
+    // Extract required parameters from the link
+    NSArray<NSString*> *pathParams;
+    NSMutableDictionary *queryParams;
+    [self parseUniversalLinkFragment:webURL.absoluteString outPathParams:&pathParams outQueryParams:&queryParams];
+
+    UniversalLink *newLink = [[UniversalLink alloc] initWithUrl:webURL pathParams:pathParams queryParams:queryParams];
+    if (![_lastHandledUniversalLink isEqual:newLink])
+    {
+        _lastHandledUniversalLink = [[UniversalLink alloc] initWithUrl:webURL pathParams:pathParams queryParams:queryParams];
+        //  notify this change
+        [[NSNotificationCenter defaultCenter] postNotificationName:AppDelegateUniversalLinkDidChangeNotification object:nil];
+    }
+
     if ([webURL.path hasPrefix:@"/config"])
     {
         return [self handleServerProvionningLink:webURL];
@@ -2047,8 +2063,39 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
     NSString *validateEmailSubmitTokenAPIPathV1 = [NSString stringWithFormat:@"/%@/%@", kMXIdentityAPIPrefixPathV1, validateEmailSubmitTokenPath];
     NSString *validateEmailSubmitTokenAPIPathV2 = [NSString stringWithFormat:@"/%@/%@", kMXIdentityAPIPrefixPathV2, validateEmailSubmitTokenPath];
     
-    // Manage email validation link
-    if ([webURL.path isEqualToString:validateEmailSubmitTokenAPIPathV1] || [webURL.path isEqualToString:validateEmailSubmitTokenAPIPathV2])
+    // Manage email validation links from homeserver for registration (/registration/email/submit_token)
+    // and email addition (/add_threepid/email/submit_token)
+    // They look like https://matrix.org/_matrix/client/unstable/registration/email/submit_token?token=vtQjQIZfwdoREDACTEDozrmKYSWlCXsJ&client_secret=53e679ea-oRED-ACTED-92b8-3012c49c6cfa&sid=qlBCREDACTEDEtgxD
+    if ([webURL.path hasSuffix:@"/email/submit_token"])
+    {
+        NSLog(@"[AppDelegate] handleUniversalLink: Validate link");
+        
+        // We just need to ping the link.
+        // The app should be in the registration flow at the "waiting for email validation" polling state. The server
+        // will indicate the email is validated through this polling API. Then, the app will go to the next flow step.
+        NSURLSessionConfiguration *conf = [NSURLSessionConfiguration defaultSessionConfiguration];
+        NSURLSession *urlSession = [NSURLSession sessionWithConfiguration:conf];
+        
+        NSURLSessionDataTask * task = [urlSession dataTaskWithURL:webURL completionHandler:^(NSData * _Nullable data, NSURLResponse * _Nullable response, NSError * _Nullable error) {
+            
+            NSLog(@"[AppDelegate] handleUniversalLink: Link validation response: %@\nData: %@", response,
+                  [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding]);
+            
+            if (error)
+            {
+                NSLog(@"[AppDelegate] handleUniversalLink: Link validation error: %@", error);
+                [self showErrorAsAlert:error];
+            }
+        }];
+        
+        [task resume];
+        
+        return YES;
+    }
+    
+    // Manage email validation link from Identity Server v1 or v2
+    else if ([webURL.path isEqualToString:validateEmailSubmitTokenAPIPathV1]
+             || [webURL.path isEqualToString:validateEmailSubmitTokenAPIPathV2])
     {
         // Validate the email on the passed identity server
         NSString *identityServer = [NSString stringWithFormat:@"%@://%@", webURL.scheme, webURL.host];
@@ -2066,12 +2113,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
         }
         
         MXIdentityService *identityService = [[MXIdentityService alloc] initWithIdentityServer:identityServer accessToken:nil andHomeserverRestClient:homeserverRestClient];
-        
-        // Extract required parameters from the link
-        NSArray<NSString*> *pathParams;
-        NSMutableDictionary *queryParams;
-        [self parseUniversalLinkFragment:webURL.absoluteString outPathParams:&pathParams outQueryParams:&queryParams];
-        
+
         NSString *clientSecret = queryParams[@"client_secret"];
         NSString *sid = queryParams[@"sid"];
         
@@ -4803,7 +4845,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
 {
     BOOL presented = NO;
     
-    if (!keyVerificationCoordinatorBridgePresenter)
+    if (!keyVerificationCoordinatorBridgePresenter.isPresenting)
     {
         NSLog(@"[AppDelegate] presentIncomingKeyVerificationRequest");
         
@@ -4827,7 +4869,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
     NSLog(@"[AppDelegate][MXKeyVerification] presentIncomingKeyVerification: %@", transaction);
 
     BOOL presented = NO;
-    if (!keyVerificationCoordinatorBridgePresenter)
+    if (!keyVerificationCoordinatorBridgePresenter.isPresenting)
     {
         keyVerificationCoordinatorBridgePresenter = [[KeyVerificationCoordinatorBridgePresenter alloc] initWithSession:mxSession];
         keyVerificationCoordinatorBridgePresenter.delegate = self;
@@ -4848,7 +4890,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
     NSLog(@"[AppDelegate][MXKeyVerification] presentUserVerificationForRoomMember: %@", roomMember);
     
     BOOL presented = NO;
-    if (!keyVerificationCoordinatorBridgePresenter)
+    if (!keyVerificationCoordinatorBridgePresenter.isPresenting)
     {
         keyVerificationCoordinatorBridgePresenter = [[KeyVerificationCoordinatorBridgePresenter alloc] initWithSession:mxSession];
         keyVerificationCoordinatorBridgePresenter.delegate = self;
@@ -4869,7 +4911,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
     NSLog(@"[AppDelegate][MXKeyVerification] presentSelfVerificationForOtherDeviceId: %@", deviceId);
     
     BOOL presented = NO;
-    if (!keyVerificationCoordinatorBridgePresenter)
+    if (!keyVerificationCoordinatorBridgePresenter.isPresenting)
     {
         keyVerificationCoordinatorBridgePresenter = [[KeyVerificationCoordinatorBridgePresenter alloc] initWithSession:mxSession];
         keyVerificationCoordinatorBridgePresenter.delegate = self;
@@ -4965,6 +5007,12 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
             {
                 // Self verification
                 NSLog(@"[AppDelegate][KeyVerification] keyVerificationNewRequestNotification: Self verification from %@", keyVerificationByToDeviceRequest.otherDevice);
+                
+                if (!self.handleSelfVerificationRequest)
+                {
+                    NSLog(@"[AppDelegate][KeyVerification] keyVerificationNewRequestNotification: Self verification handled elsewhere");
+                    return;
+                }
                       
                 NSString *myUserId = keyVerificationByToDeviceRequest.otherUser;
                 MXKAccount *account = [[MXKAccountManager sharedManager] accountForUserId:myUserId];
@@ -5006,6 +5054,12 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
                                                senderId:(NSString*)senderId
                                                 request:(MXKeyVerificationRequest*)keyVerificationRequest
 {
+    if (keyVerificationRequest.state != MXKeyVerificationRequestStatePending)
+    {
+        NSLog(@"[AppDelegate] presentNewKeyVerificationRequest: Request already accepted. Do not display it");
+        return;
+    }
+    
     if (self.incomingKeyVerificationRequestAlertController)
     {
         [self.incomingKeyVerificationRequestAlertController dismissViewControllerAnimated:NO completion:nil];
@@ -5021,7 +5075,17 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
     {
         senderInfo = senderId;
     }
+
     
+    __block id observer;
+    void (^removeObserver)(void) = ^() {
+        if (observer)
+        {
+            [[NSNotificationCenter defaultCenter] removeObserver:observer];
+            observer = nil;
+        }
+    };
+
     
     UIAlertController *alertController = [UIAlertController alertControllerWithTitle:NSLocalizedStringFromTable(@"key_verification_tile_request_incoming_title", @"Vector", nil)
                                                                                              message:senderInfo
@@ -5031,6 +5095,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
                                                                                            style:UIAlertActionStyleDefault
                                                                                          handler:^(UIAlertAction * action)
                                                                    {
+                                                                       removeObserver();
                                                                        [self presentIncomingKeyVerificationRequest:keyVerificationRequest inSession:session];
                                                                    }]];
     
@@ -5038,6 +5103,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
                                                                                            style:UIAlertActionStyleDestructive
                                                                                          handler:^(UIAlertAction * action)
                                                                    {
+                                                                       removeObserver();
                                                                        [keyVerificationRequest cancelWithCancelCode:MXTransactionCancelCode.user success:^{
                                                                            
                                                                        } failure:^(NSError * _Nonnull error) {
@@ -5049,11 +5115,26 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
                                                                                            style:UIAlertActionStyleCancel
                                                                                          handler:^(UIAlertAction * action)
                                                                    {
+                                                                       removeObserver();
                                                                    }]];
     
     [self presentViewController:alertController animated:YES completion:nil];
-    
     self.incomingKeyVerificationRequestAlertController = alertController;
+    
+    observer = [[NSNotificationCenter defaultCenter] addObserverForName:MXKeyVerificationRequestDidChangeNotification
+                                                                 object:keyVerificationRequest
+                                                                  queue:[NSOperationQueue mainQueue]
+                                                             usingBlock:^(NSNotification * _Nonnull note)
+                {
+                    if (keyVerificationRequest.state != MXKeyVerificationRequestStatePending)
+                    {
+                        if (self.incomingKeyVerificationRequestAlertController == alertController)
+                        {
+                            [self.incomingKeyVerificationRequestAlertController dismissViewControllerAnimated:NO completion:nil];
+                            removeObserver();
+                        }
+                    }
+                }];
 }
 
 #pragma mark - New Sign In
@@ -5127,14 +5208,14 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
                                                                    message:alertMessage
                                                             preferredStyle:UIAlertControllerStyleAlert];
     
-    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"continue", @"Vector", nil)
+    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"device_verification_self_verify_alert_validate_action", @"Vector", nil)
                                               style:UIAlertActionStyleDefault
                                             handler:^(UIAlertAction * action) {
                                                 
                                                 [self presentSelfVerificationForOtherDeviceId:device.deviceId inSession:session];
                                             }]];
     
-    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"device_verification_self_verify_alert_cancel_action", @"Vector", nil)
+    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"later", @"Vector", nil)
                                                style:UIAlertActionStyleCancel
                                              handler:nil]];
      
@@ -5150,7 +5231,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
     NSLog(@"[AppDelegate][MXKeyVerification] presentCompleteSecurityForSession");
     
     BOOL presented = NO;
-    if (!keyVerificationCoordinatorBridgePresenter)
+    if (!keyVerificationCoordinatorBridgePresenter.isPresenting)
     {
         keyVerificationCoordinatorBridgePresenter = [[KeyVerificationCoordinatorBridgePresenter alloc] initWithSession:mxSession];
         keyVerificationCoordinatorBridgePresenter.delegate = self;